+++ traffic_ops/ort/atstccfg/cfgfile/servercachedotconfig.go
@@ -21,57 +21,26 @@ package cfgfile
 
 import (
 	"errors"
-	"strconv"
 	"strings"
 
 	"github.com/apache/trafficcontrol/lib/go-atscfg"
 	"github.com/apache/trafficcontrol/lib/go-tc"
 	"github.com/apache/trafficcontrol/traffic_ops/ort/atstccfg/config"
-	"github.com/apache/trafficcontrol/traffic_ops/ort/atstccfg/toreq"
 )
 
 const ServerCacheDotConfigIncludeInactiveDSes = false // TODO move to lib/go-atscfg
 
-func GetConfigFileServerCacheDotConfig(cfg config.TCCfg, serverNameOrID string) (string, error) {
+func GetConfigFileServerCacheDotConfig(toData *config.TOData) (string, string, error) {
 	// TODO TOAPI add /servers?cdn=1 query param
-	servers, err := toreq.GetServers(cfg)
-	if err != nil {
-		return "", errors.New("getting servers: " + err.Error())
-	}
 
-	server := tc.Server{ID: atscfg.InvalidID}
-	if serverID, err := strconv.Atoi(serverNameOrID); err == nil {
-		for _, toServer := range servers {
-			if toServer.ID == serverID {
-				server = toServer
-				break
-			}
-		}
-	} else {
-		serverName := serverNameOrID
-		for _, toServer := range servers {
-			if toServer.HostName == serverName {
-				server = toServer
-				break
-			}
-		}
-	}
-	if server.ID == atscfg.InvalidID {
-		return "", errors.New("server '" + serverNameOrID + " not found in servers")
-	}
-
-	if !strings.HasPrefix(string(server.Type), tc.MidTypePrefix) {
+	// TODO remove this, we generated the scope, we know it's right? Or should we have an extra safety check?
+	if !strings.HasPrefix(string(toData.Server.Type), tc.MidTypePrefix) {
 		// emulates Perl
-		return "", errors.New("Error - incorrect file scope for route used.  Please use the profiles route.")
-	}
-
-	dses, err := toreq.GetCDNDeliveryServices(cfg, server.CDNID)
-	if err != nil {
-		return "", errors.New("getting delivery services: " + err.Error())
+		return "", "", errors.New("Error - incorrect file scope for route used.  Please use the profiles route.")
 	}
 
 	dsData := map[tc.DeliveryServiceName]atscfg.ServerCacheConfigDS{}
-	for _, ds := range dses {
+	for _, ds := range toData.DeliveryServices {
 		if ds.XMLID == nil || ds.Active == nil || ds.OrgServerFQDN == nil || ds.Type == nil {
 			// TODO orgserverfqdn is nil for some DSes - MSO? Verify.
 			continue
@@ -83,13 +52,7 @@ func GetConfigFileServerCacheDotConfig(cfg config.TCCfg, serverNameOrID string) 
 		dsData[tc.DeliveryServiceName(*ds.XMLID)] = atscfg.ServerCacheConfigDS{OrgServerFQDN: *ds.OrgServerFQDN, Type: *ds.Type}
 	}
 
-	serverName := tc.CacheName(server.HostName)
+	serverName := tc.CacheName(toData.Server.HostName)
 
-	toToolName, toURL, err := toreq.GetTOToolNameAndURLFromTO(cfg)
-	if err != nil {
-		return "", errors.New("getting global parameters: " + err.Error())
-	}
-
-	txt := atscfg.MakeServerCacheDotConfig(serverName, toToolName, toURL, dsData)
-	return txt, nil
+	return atscfg.MakeServerCacheDotConfig(serverName, toData.TOToolName, toData.TOURL, dsData), atscfg.ContentTypeCacheDotConfig, nil
 }